refactor: resources to extensions
