+++ pixi.toml
@@ -6,7 +6,7 @@ platforms = ["linux-64", "osx-64", "osx-arm64", "linux-aarch64"]
 [tasks]
 python-generate-antlr = { cmd = "cd python/substrait-antlr && ./generate_antlr.sh", description = "Generate ANTLR Python parsers" }
 python-generate-protobuf = { cmd = "cd python/substrait-protobuf && ./generate_protobuf.sh", description = "Generate protobuf Python code" }
-python-generate-resources = { cmd = "cd python/substrait-resources && ./generate_resources.sh", description = "Package Substrait resource files and YAML schema-based code" }
+python-generate-extensions = { cmd = "cd python/substrait-extensions && ./generate_extensions.sh", description = "Package Substrait extensions files" }
 
 [dependencies]
 antlr = ">=4.13.2,<5"
